Remodel Procedure Activity Act/performer in order to remove all custom constraints and fix invalid validation results. See: artf3573 : Provide a model based fix for ProcedureActPerformer, similar to ProcedureActivityObservation and SITE-653
Note:
There are extra unrelated changes being committed:
Instead of \n - on windows the same is represented as \r\n after a source cleanup.
I have decided to commit the files instead of un-checking them all as I will be the regularly committing future changes. Therefore, it would make the most sense for the format to match my machine so I can see the real differences and not have this problem for all future commits.