Generated code for: Fixed/removed extension required in ProgressNote/DocumentationOf/ServiceEvent2. The error was introduced due to the '2' in the ServiceEvent class confusing the script which was designed to add the extension to all R2 V2 templates. The validateProgressNoteDocumentationOfServiceEvent2TemplateId JUnit passes now.
Note: The reason everything appears to be modified is due to differences in OS/computers after formatting the source. Since I do most of the commits for R2 (for the time being) it will make the most sense to keep the Windows formatting so that the differences can be seen/verified in the future.