Revert errata 836 so ReactionObservation2/value term matches R1.1 impl.
-Errata 836 says "3.90	Reaction Observation (V2) 
[observation: identifier
urn:hl7ii:2.16.840.1.113883.10.20.22.4.9:2014-06-09 (open)]
SHALL contain exactly one [1..1] value with @xsi:type="CD", which SHALL
be selected from ValueSet Reaction Severity [new OID] STATIC (CONF:NEW)"
But we are reverting to the R1.1 implementation (selects from the
Problem Value Set). This is because there is errata in the errata
comment (quoted) as it was supposed to describe a change to Severity
Observation instead.